# backend/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v1
components:
  schemas:
...
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
        name:
          type: string
        company:
          type: string
        price:
          type: number
          format: double
        description:
          type: string
        stock:
          type: integer
          format: int32
        isAccepted:
          type: boolean
        date:
          type: string
          format: date-time
        views_count:
          type: integer
          format: int32
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        imageBase64:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    SearchProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        type:
          type: string
        imageBase64:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
servers:
  - url: https://localhost:8443
    description: Generated server url
paths:
  /api/v1/users/{id}/image:
    get:
      tags:
        - api-user-controller
      summary: Get user profile image
      description: Retrieves the profile image for the specified user
      operationId: getUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User image retrieved successfully
          content:
            image/*:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized - Missing or invalid authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found or has no profile image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags:
        - api-user-controller
      summary: Replace user profile image
      description: Uploads a new profile image replacing any existing one
      operationId: replaceUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: The image file to upload (JPG/PNG)
              required:
                - imageFile
      responses:
        "200":
          description: Image replaced successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  imageUrl:
                    type: string
        "400":
          description: Invalid file format or missing file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags:
        - api-user-controller
      summary: Upload user profile image
      description: Uploads a profile image (requires no existing image)
      operationId: uploadUserImage
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric ID of the user
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: The image file to upload (JPG/PNG)
              required:
                - imageFile
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  imageUrl:
                    type: string
        "400":
          description: User already has a profile image (use PUT instead)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
      - api-user-controller
      summary: Delete user image
      description: Delete a user's image.
      operationId: deleteProductImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User image deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Image doesn't exist
          content:
            '*/*':
              schema:
                type: object
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                type: object
...
  /api/v1/products/{id}:
    get:
      tags:
      - api-product-controller
      summary: Get product by ID
      description: Retrieve a product by its unique ID.
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    put:
      tags:
      - api-product-controller
      summary: Update a product
      description: Update the details of an existing product.
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductDTO"
        required: true
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
    delete:
      tags:
      - api-product-controller
      summary: Delete a product
      description: Delete a product by its unique ID.
      operationId: deleteProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Product deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
        "404":
          description: Product not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/products/addViewsCount:
    put:
      summary: Increment views count
      tags:
      - api-product-controller
      operationId: addViewsCount
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/accept:
    put:
      summary: Accept product
      tags:
      - api-product-controller
      operationId: acceptProduct
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ProductDTO"
...
  /api/v1/products/type:
    get:
      tags:
      - api-product-controller
      summary: Get products by type
      description: Retrieve products filtered by type.
      operationId: getProductsByType
      parameters:
      - name: type
        in: query
        required: true
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/type/all:
    get:
      tags:
      - api-product-controller
      summary: Get products by type
      description: Retrieve products filtered by type.
      operationId: getAllProductsByType
      parameters:
      - name: type
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of products retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDTO"
  /api/v1/products/search:
    get:
      summary: Search products
      tags:
      - api-product-controller
      operationId: searchProducts
      parameters:
      - name: search_text
        in: query
        required: false
        schema:
          type: string
      - name: type
        in: query
        required: false
        schema:
          type: string
          default: all
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchProductDTO"
...